+++ bin/Debug/dep/class_classList.cpp.d
@@ -1,0 +1,6 @@
+bin/Debug/obj/class_classList.cpp.o: src/class_classList.cpp src/header.h \
+ src/struct_and_dependencies.h
+
+src/header.h:
+
+src/struct_and_dependencies.h:
