+++ bin/Debug/dep/course_courseList.cpp.d
@@ -1,0 +1,6 @@
+bin/Debug/obj/course_courseList.cpp.o: src/course_courseList.cpp \
+ src/header.h src/struct_and_dependencies.h
+
+src/header.h:
+
+src/struct_and_dependencies.h:
