+++ bin/Debug/dep/semester_semesterList.cpp.d
@@ -1,0 +1,6 @@
+bin/Debug/obj/semester_semesterList.cpp.o: src/semester_semesterList.cpp \
+ src/header.h src/struct_and_dependencies.h
+
+src/header.h:
+
+src/struct_and_dependencies.h:
