+++ bin/Debug/dep/student_studentList.cpp.d
@@ -1,0 +1,6 @@
+bin/Debug/obj/student_studentList.cpp.o: src/student_studentList.cpp \
+ src/header.h src/struct_and_dependencies.h
+
+src/header.h:
+
+src/struct_and_dependencies.h:
