+++ src/sample_student_output.csv
@@ -1,0 +1,4 @@
+2,21834208,Van,Nguyen,3/2/2000,90889839,TrTRTR,
+3,21039491,Hung,Pham,2/2/2003,23908029,,
+4,23284030,Khanh,Do,3/11/2000,23891203,,
+5,27329340,Phuong,Bich,3/9/2003,21318239,,
