--- student_sample_data.csv
@@ -1,5 +1,0 @@
-1,18125030,Triet,Phan,0,4/9/2000,02585933,TrTrTr,18CTT3
-2,21834208,Van,Nguyen,1,3/2/2000,90889839,,
-3,21039491,Hung,Pham,0,2/2/2003,23908029,,
-4,23284030,Khanh,Do,0,3/11/2000,23891203,,
-5,27329340,Phuong,Bich,1,3/9/2003,21318239,,
